+++ page-templates/about.php
@@ -11,6 +11,6 @@ get_header();
 get_template_part('template-parts/pages/about/page-header');
 get_template_part('template-parts/pages/about/image-text');
 get_template_part('template-parts/pages/about/featured-gallery');
-get_template_part('template-parts/components/cta-promo');
+get_template_part('template-parts/pages/global/cta-promo');
 
 get_footer();