+++ page-templates/exhibit.php
@@ -8,7 +8,7 @@
 
 get_header();
 
-get_template_part('template-parts/pages/page-header-exhibit');
-get_template_part('template-parts/components/exhibit-with-us');
+get_template_part('template-parts/pages/exhibit/page-header-exhibit');
+get_template_part('template-parts/pages/exhibit/exhibit-with-us');
 
 get_footer();