+++ page-templates/homepage.php
@@ -8,11 +8,11 @@
 
 get_header();
 
-get_template_part('template-parts/components/home-hero-slider');
-get_template_part('template-parts/components/interactive-select');
-get_template_part('template-parts/components/shows-cards-grid');
-get_template_part('template-parts/components/subscribe');
-get_template_part('template-parts/components/featured-gallery');
-get_template_part('template-parts/components/cta-promo');
+get_template_part('template-parts/pages/home/home-hero-slider');
+get_template_part('template-parts/pages/home/interactive-select');
+get_template_part('template-parts/pages/home/shows-cards-grid');
+get_template_part('template-parts/pages/global/subscribe');
+get_template_part('template-parts/pages/global/featured-gallery');
+get_template_part('template-parts/pages/global/cta-promo');
 
 get_footer();